+++ services/storage/rpc_service.go
@@ -8,13 +8,23 @@ import (
 	"strings"
 
 	"github.com/gogo/protobuf/types"
+	"github.com/influxdata/influxdb/pkg/metrics"
 	"github.com/influxdata/influxdb/tsdb"
+	"github.com/influxdata/influxdb/tsdb/engine/tsm1"
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 	"go.uber.org/zap"
 )
 
 //go:generate protoc -I$GOPATH/src -I. --plugin=protoc-gen-yarpc=$GOPATH/bin/protoc-gen-yarpc --yarpc_out=Mgoogle/protobuf/empty.proto=github.com/gogo/protobuf/types:. --gogofaster_out=Mgoogle/protobuf/empty.proto=github.com/gogo/protobuf/types:. storage.proto predicate.proto
 //go:generate tmpl -data=@batch_cursor.gen.go.tmpldata batch_cursor.gen.go.tmpl
 //go:generate tmpl -data=@batch_cursor.gen.go.tmpldata response_writer.gen.go.tmpl
+
+const (
+	batchSize  = 1000
+	frameCount = 50
+	writeSize  = 64 << 10 // 64k
+)
 
 type rpcService struct {
 	loggingEnabled bool
@@ -31,26 +41,58 @@ func (r *rpcService) Hints(context.Context, *types.Empty) (*HintsResponse, error
 	return nil, errors.New("not implemented")
 }
 
-const (
-	batchSize  = 1000
-	frameCount = 50
-	writeSize  = 64 << 10 // 64k
-)
-
 func (r *rpcService) Read(req *ReadRequest, stream Storage_ReadServer) error {
 	// TODO(sgc): implement frameWriter that handles the details of streaming frames
+	var err error
+	var wire opentracing.SpanContext
+
+	if len(req.Trace) > 0 {
+		wire, err = opentracing.GlobalTracer().Extract(opentracing.TextMap, opentracing.TextMapCarrier(req.Trace))
+		if err != nil {
+			// TODO(sgc): log it?
+		}
+	}
+
+	span := opentracing.StartSpan("storage.read", ext.RPCServerOption(wire))
+	defer span.Finish()
+
+	ext.DBInstance.Set(span, req.Database)
+
+	// TODO(sgc): use yarpc stream.Context() once implemented
+	ctx := context.Background()
+	ctx = opentracing.ContextWithSpan(ctx, span)
+	// TODO(sgc): this should be available via a generic API, such as tsdb.Store
+	ctx = tsm1.NewContextWithMetricsGroup(ctx)
+
+	var agg Aggregate_AggregateType
+	if req.Aggregate != nil {
+		agg = req.Aggregate.Type
+	}
+	pred := truncateString(PredicateToExprString(req.Predicate))
+	groupKeys := truncateString(strings.Join(req.Grouping, ","))
+	span.
+		SetTag("predicate", pred).
+		SetTag("series_limit", req.SeriesLimit).
+		SetTag("series_offset", req.SeriesOffset).
+		SetTag("points_limit", req.PointsLimit).
+		SetTag("start", req.TimestampRange.Start).
+		SetTag("end", req.TimestampRange.End).
+		SetTag("desc", req.Descending).
+		SetTag("group_keys", groupKeys).
+		SetTag("aggregate", agg.String())
 
 	if r.loggingEnabled {
 		r.Logger.Info("request",
 			zap.String("database", req.Database),
-			zap.String("predicate", PredicateToExprString(req.Predicate)),
+			zap.String("predicate", pred),
 			zap.Uint64("series_limit", req.SeriesLimit),
 			zap.Uint64("series_offset", req.SeriesOffset),
 			zap.Uint64("points_limit", req.PointsLimit),
 			zap.Int64("start", req.TimestampRange.Start),
 			zap.Int64("end", req.TimestampRange.End),
 			zap.Bool("desc", req.Descending),
-			zap.String("grouping", strings.Join(req.Grouping, ",")),
+			zap.String("group_keys", groupKeys),
+			zap.String("aggregate", agg.String()),
 		)
 	}
 
@@ -58,8 +100,7 @@ func (r *rpcService) Read(req *ReadRequest, stream Storage_ReadServer) error {
 		req.PointsLimit = math.MaxUint64
 	}
 
-	// TODO(sgc): use yarpc stream.Context() once implemented
-	rs, err := r.Store.Read(context.Background(), req)
+	rs, err := r.Store.Read(ctx, req)
 	if err != nil {
 		r.Logger.Error("Store.Read failed", zap.Error(err))
 		return err
@@ -106,6 +147,15 @@ func (r *rpcService) Read(req *ReadRequest, stream Storage_ReadServer) error {
 	}
 
 	w.flushFrames()
+
+	span.SetTag("num_values", w.vc)
+	grp := tsm1.MetricsGroupFromContext(ctx)
+	grp.ForEach(func(v metrics.Metric) {
+		switch m := v.(type) {
+		case *metrics.Counter:
+			span.SetTag(m.Name(), m.Value())
+		}
+	})
 
 	return nil
 }